requirements.txt
numpy
Pillow
requests
gradio
huggingface_hub<1.0
PyYAML
